[FLINK-14699][type] Move ClosureCleaner to flink-core
